# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chatbot Templates API",
    "version": "2.0.0"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Root",
        "description": "Root endpoint with API information.",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "description": "Health check endpoint.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/chat": {
      "post": {
        "summary": "Chat",
        "description": "Chat endpoint that generates responses using an LLM.\n\nThe response is generated based on:\n- The system prompt loaded from prompts/rag_system_prompt.txt\n- The user's message\n- Optional context provided in the request\n\nArgs:\n    request: ChatRequest with message and optional context\n\nReturns:\n    ChatResponse with the generated response, system prompt, and context flag\n\nRaises:\n    HTTPException: If message is empty or LLM generation fails",
        "operationId": "chat_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chat-with-retrieval": {
      "post": {
        "summary": "Chat With Retrieval",
        "description": "Chat endpoint with automatic context retrieval.\n\nThis endpoint:\n1. Takes the user's message\n2. Retrieves relevant context from the vector database (currently a stub)\n3. Generates a response using the LLM with the retrieved context\n\nArgs:\n    request: ChatRequest with message (context field is ignored)\n\nReturns:\n    ChatResponse with the generated response, system prompt, and context flag\n\nRaises:\n    HTTPException: If message is empty or processing fails",
        "operationId": "chat_with_retrieval_chat_with_retrieval_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/support-bot/query": {
      "post": {
        "summary": "Support Bot Query",
        "description": "FAQ-backed support assistant with tenant-aware RAG over docs/faq collections.",
        "operationId": "support_bot_query_support_bot_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupportBotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportBotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest",
        "description": "Request model for chat endpoint."
      },
      "ChatResponse": {
        "properties": {
          "response": {
            "type": "string",
            "title": "Response"
          },
          "system_prompt": {
            "type": "string",
            "title": "System Prompt"
          },
          "context_used": {
            "type": "boolean",
            "title": "Context Used"
          }
        },
        "type": "object",
        "required": [
          "response",
          "system_prompt",
          "context_used"
        ],
        "title": "ChatResponse",
        "description": "Response model for chat endpoint."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "SupportBotRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "message"
        ],
        "title": "SupportBotRequest",
        "description": "Request payload for the support bot endpoint."
      },
      "SupportBotResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "fallback_used": {
            "type": "boolean",
            "title": "Fallback Used"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          },
          "retrieved_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieved Context"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "answer",
          "fallback_used",
          "sources",
          "tenant_id"
        ],
        "title": "SupportBotResponse",
        "description": "Response payload for the support bot endpoint."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}